setting up IndexService in performance test

+++ jbpm-human-task/jbpm-human-task-audit/src/main/java/org/jbpm/services/task/audit/service/TaskAuditService.java
@@ -37,8 +37,8 @@ import org.kie.internal.task.api.model.TaskEvent;
 public interface TaskAuditService {
     void setTaskService(TaskService taskService);
     
-    List<TaskEvent> getAllTaskEvents(long taskId, int offset, int count);
-    QueryResult<TaskEvent> getTaskEvents(long offset, int count, QueryComparator<TaskEvent> comparator, Filter<TaskEvent, ?>... filters);
+    List<TaskEvent> getAllTaskEvents(int taskId, int offset, int count);
+    QueryResult<TaskEvent> getTaskEvents(int offset, int count, QueryComparator<TaskEvent> comparator, Filter<TaskEvent, ?>... filters);
 
 
     List<UserAuditTask> getAllUserAuditTasksAdmin(int offset, int count);
@@ -47,7 +47,7 @@ public interface TaskAuditService {
     List<UserAuditTask> getAllUserAuditTasksByDueDate(String userId, Date dueDate, int offset, int count);
     List<UserAuditTask> getAllUserAuditTasksByStatusByDueDate(String userId, List<String> statuses, Date dueDate, int offset, int count);
     List<UserAuditTask> getAllUserAuditTasksByStatusByDueDateOptional(String userId, List<String> statuses, Date dueDate, int offset, int count);
-    QueryResult<UserAuditTask> getUserAuditTasks(long offset, int count, QueryComparator<UserAuditTask> comparator, Filter<UserAuditTask, ?>... filters);
+    QueryResult<UserAuditTask> getUserAuditTasks(int offset, int count, QueryComparator<UserAuditTask> comparator, Filter<UserAuditTask, ?>... filters);
 
     List<GroupAuditTask> getAllGroupAuditTasksAdmin(int offset, int count);
     List<GroupAuditTask> getAllGroupAuditTasks(String groupIds, int offset, int count);
@@ -55,13 +55,13 @@ public interface TaskAuditService {
     List<GroupAuditTask> getAllGroupAuditTasksByDueDate(String groupIds, Date dueDate, int offset, int count);
     List<GroupAuditTask> getAllGroupAuditTasksByStatusByDueDate(String groupIds, List<String> statuses, Date dueDate, int offset, int count);
     List<GroupAuditTask> getAllGroupAuditTasksByStatusByDueDateOptional(String groupIds, List<String> statuses, Date dueDate, int offset, int count);
-    QueryResult<GroupAuditTask> getGroupAuditTasks(long offset, int count, QueryComparator<GroupAuditTask> comparator, Filter<GroupAuditTask, ?>... filters);
+    QueryResult<GroupAuditTask> getGroupAuditTasks(int offset, int count, QueryComparator<GroupAuditTask> comparator, Filter<GroupAuditTask, ?>... filters);
 
 
 
     List<HistoryAuditTask> getAllHistoryAuditTasks( int offset, int count);
     List<HistoryAuditTask> getAllHistoryAuditTasksByUser(String userId, int offset, int count);
-    QueryResult<HistoryAuditTask> getHistoryAuditTasks(long offset, int count, QueryComparator<HistoryAuditTask> comparator, Filter<HistoryAuditTask, ?>... filters);
+    QueryResult<HistoryAuditTask> getHistoryAuditTasks(int offset, int count, QueryComparator<HistoryAuditTask> comparator, Filter<HistoryAuditTask, ?>... filters);
